+++ tests/test_bayes.py
@@ -23,6 +23,11 @@ def test_bayesnet():
     assert out == 1
 
 
+def test_bayesnet_noop():
+    out = bayesnet()
+    assert out is None
+
+
 def test_bayesnet_conditional():
     def define_event():
         a = sample(discrete([1, 2]))
@@ -216,11 +221,11 @@ def test_bayesnet_insufficent_samples_error():
 def cachefile():
     cachefile = 'testcache'
     yield cachefile
-    os.remove(cachefile + '.sqcache.pkl')
+    os.remove(cachefile + '.sqcache.json')
 
 
 def test_bayesnet_cachefile(cachefile):
-    assert not os.path.exists(cachefile + '.sqcache.pkl')
+    assert not os.path.exists(cachefile + '.sqcache.json')
 
     def define_event():
         return {'a': 1, 'b': 2}
@@ -234,7 +239,7 @@ def test_bayesnet_cachefile(cachefile):
                    find=lambda e: e['a'],
                    raw=True,
                    n=100)
-    assert os.path.exists(cachefile + '.sqcache.pkl')
+    assert os.path.exists(cachefile + '.sqcache.json')
     assert set(out) == set([1])
 
     out = bayesnet(define_event,
@@ -242,7 +247,7 @@ def test_bayesnet_cachefile(cachefile):
                    conditional_on=lambda e: e['b'] == 2,
                    raw=True,
                    n=100)
-    assert os.path.exists(cachefile + '.sqcache.pkl')
+    assert os.path.exists(cachefile + '.sqcache.json')
     assert set(out) == set([1])
 
     def define_event():
@@ -265,7 +270,7 @@ def test_bayesnet_cachefile(cachefile):
 
 
 def test_bayesnet_cachefile_primary(cachefile):
-    assert not os.path.exists(cachefile + '.sqcache.pkl')
+    assert not os.path.exists(cachefile + '.sqcache.json')
 
     from ..squigglepy.bayes import _squigglepy_internal_bayesnet_caches
     n_caches = len(_squigglepy_internal_bayesnet_caches)
@@ -278,7 +283,7 @@ def test_bayesnet_cachefile_primary(cachefile):
     from ..squigglepy.bayes import _squigglepy_internal_bayesnet_caches
     n_caches2 = len(_squigglepy_internal_bayesnet_caches)
     assert n_caches2 == n_caches + 1
-    assert not os.path.exists(cachefile + '.sqcache.pkl')
+    assert not os.path.exists(cachefile + '.sqcache.json')
 
     def define_event2():
         return {'a': 2, 'b': 3}
@@ -292,7 +297,7 @@ def test_bayesnet_cachefile_primary(cachefile):
     from ..squigglepy.bayes import _squigglepy_internal_bayesnet_caches
     n_caches3 = len(_squigglepy_internal_bayesnet_caches)
     assert n_caches3 == n_caches2
-    assert os.path.exists(cachefile + '.sqcache.pkl')
+    assert os.path.exists(cachefile + '.sqcache.json')
 
     out = bayesnet(define_event,
                    find=lambda e: e['a'],
@@ -319,11 +324,11 @@ def test_bayesnet_cachefile_primary(cachefile):
     from ..squigglepy.bayes import _squigglepy_internal_bayesnet_caches
     n_caches4 = len(_squigglepy_internal_bayesnet_caches)
     assert n_caches4 == n_caches2
-    assert os.path.exists(cachefile + '.sqcache.pkl')
+    assert os.path.exists(cachefile + '.sqcache.json')
 
 
 def test_bayesnet_cachefile_will_also_memcache(cachefile):
-    assert not os.path.exists(cachefile + '.sqcache.pkl')
+    assert not os.path.exists(cachefile + '.sqcache.json')
     from ..squigglepy.bayes import _squigglepy_internal_bayesnet_caches
     n_caches = len(_squigglepy_internal_bayesnet_caches)
 
@@ -337,7 +342,7 @@ def test_bayesnet_cachefile_will_also_memcache(cachefile):
                    raw=True,
                    n=100)
 
-    assert os.path.exists(cachefile + '.sqcache.pkl')
+    assert os.path.exists(cachefile + '.sqcache.json')
     assert set(out) == set([1])
     from ..squigglepy.bayes import _squigglepy_internal_bayesnet_caches
     n_caches2 = len(_squigglepy_internal_bayesnet_caches)
@@ -350,7 +355,7 @@ def test_bayesnet_cachefile_will_also_memcache(cachefile):
                    raw=True,
                    n=100)
 
-    assert os.path.exists(cachefile + '.sqcache.pkl')
+    assert os.path.exists(cachefile + '.sqcache.json')
     assert set(out) == set([1])
     from ..squigglepy.bayes import _squigglepy_internal_bayesnet_caches
     n_caches3 = len(_squigglepy_internal_bayesnet_caches)
@@ -358,7 +363,7 @@ def test_bayesnet_cachefile_will_also_memcache(cachefile):
 
 
 def test_bayesnet_cachefile_insufficent_samples_error(cachefile):
-    assert not os.path.exists(cachefile + '.sqcache.pkl')
+    assert not os.path.exists(cachefile + '.sqcache.json')
 
     def define_event():
         return {'a': 1, 'b': 2}
@@ -367,7 +372,7 @@ def test_bayesnet_cachefile_insufficent_samples_error(cachefile):
              find=lambda e: e['a'],
              dump_cache_file=cachefile,
              n=100)
-    assert os.path.exists(cachefile + '.sqcache.pkl')
+    assert os.path.exists(cachefile + '.sqcache.json')
 
     with pytest.raises(ValueError) as excinfo:
         bayesnet(define_event,