+++ CHANGES.md
@@ -1,11 +1,13 @@
 ## v0.18
 
 * **[Breaking change]** The default `t` for t-distributions has changed from 1 to 20.
+* **[Breaking change]** `bayesnet` caching is now based on json files instead of pickle files.
 * Fixes a bug where `lclip` and/or `rclip` on `mixture` distribution were not working correctly.
 * Fixes a bug where `dist_fn` did not work with `np.vectorize` functions.
 * Fixes a bug where in-memory caching was invoked for `bayesnet` when not desired.
 * `bayesnet` and `sample` now take an argument `cores` (default 1). If greater than 1, will run the calculations on multiple cores using the pathos package.
 * `sample` results can now be cached in-memory using `memcache=True`. They can also be cached to a file -- use `dump_cache_file` to write the file and `load_cache_file` to load from the file.
+* Verbose output for `bayesnet` and `sample` is now clearer (and slightly more verbose).
 * _(Non-visible backend change)_ Weights that are set to 0 are now dropped entirely, for a potential speedup.